requirements.txt
scrapy>=2.11.0
psycopg2>=2.9.10
python-dotenv>=1.0.0
pydantic>=2.6.1
pydantic-settings>=2.1.0
